+++ doc/zmq_getsockopt.txt
@@ -464,9 +464,9 @@ Default value:: null string
 Applicable socket types:: all, when using TCP or IPC transports
 
 
-ZMQ_PRE_ALLOCATED_FD: Retrieve the pre-allocated socket file descriptor
+ZMQ_USEFD: Retrieve the pre-allocated socket file descriptor
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-The 'ZMQ_PRE_ALLOCATED_FD' option shall retrieve the pre-allocated file
+The 'ZMQ_USEFD' option shall retrieve the pre-allocated file
 descriptor that has been assigned to a ZMQ socket, if any. -1 shall be
 returned if a pre-allocated file descriptor was not set for the socket.
