+++ tests/CMakeLists.txt
@@ -90,8 +90,8 @@ if(NOT WIN32)
           test_stream_exceeds_buffer
           test_router_mandatory_hwm
           test_term_endpoint_tipc
-          test_pre_allocated_fd_ipc
-          test_pre_allocated_fd_tcp
+          test_usefd_ipc
+          test_usefd_tcp
   )
   if(HAVE_FORK)
     list(APPEND tests test_fork)
